Major change: deleted unused files and redesigned Installer
The changes are:
+ modified the manifest file in Movex.View to ask for administrator privileges
+ modified the ContextMenu module of Movex to call Movex.View and not Movex.ShortcutView
+ created a new folder inside the project called resources
+ added ServerRegistrationManagex.exe and SharpShell.dll (both version 2.2.0) in resources folder
+ modified Movex.Installer to adopt new ServerRegistrationManager to install the ContextMenu entry
+ modified Movex.Installer to install Movex.View instead of Movex.ShortcutView

From now on, the idea is to launch the Movex.View, with the selected folders and files, instead of launching
ShortcutView since this was no longer maintained. WIth this commit this module will continue to be present
because it contains the code to pass external files and folders list to the application at startup. But when this
code is transferred to Movex.View the module Movex.ShortcutView will be deleted.

Next step \ TODO:
+ copy the code from shortcutView to View
+ adapt View to load the parameters received (files and folders)
+ design the View in a way it is able to display the parameters received
+ delete ShortcutView