Completed launch of the View from the ContextualMenu
The application now, once installed, can be launched using
the ContextualMenu using the Windows Explorer doing
a clean and simple right-click on both files or folders.

Once the application is launched, the items (files or folders)
selected are shown the list overview, as usual, so the user
if changed his mind can delete items or add others!

The Send button is disabled until it add a user first!

Next Steps \ Todo:
+ delete ShortcutView
+ explain to the user why the button is disabled
+ think what the app should do when the user
try to launch it, when it is already running

Notes:
There is a folder that is not included in the git system:
the folder is the ContextMenu which is in my Documents
folder. I need to check how to solve also this thing.