Added msg for the user in ContextualMenu case
In this commit I also removed the ShortcutView
from the app, since I integrated it inside View itself.

Next Steps \ Todo:
+ improve stability and smoothness
+ think what the app should do when the user
try to launch it, when it is already running

Notes:
There is a folder that is not included in the git system:
the folder is the ContextMenu which is in my Documents
folder. I need to check how to solve also this thing.