Preparing the stuff for the major change of tomorrow commit
I looked for the locations inside the source code where was required
to make a change in order to solve the issue regarding the interrupt
button. At the moment of the trigger of the event two things have to
happen:
a) the progressControl stops the IoC.FTP.Client or IoC.FTP.Server
operation interrupting it
b) the progressControl say to its parent to close the window
c) the progressWindow (upload- or download-) has to manage
the sync primitive to wake it up and close the window.

The difficulty stays in the progressControl that hasn't to be general,
once for all, but specific for that progressWindow. And this thing
has to be modified. Anyway I wrote down some notes and
already designed where to make the modifications.