Major Fix
The transferWindow now disappear when the transfer is interrupted.
And this trigger the app to show a MessageWindow to inform the user
that the transfer has been interrupted.

Fixed bug in FTPServer. It had ip:port as "from" parameter.
So when getFrom() was invoked it returned ip:port instead of ip.
So fixed it.

Fixed MessageWindow. Now for the first time used!
Added the messageWindow to the WindowDispatcher too.

Adapted the ProgressViewModel to manage both case Upload
and Transfer thanks to a new property Type.

Also fixed graphic bug: when the transfer start the app is hidden, but the page displayed before the hide was the BrowsePage, and when
the app is resumed the page go from BrowsePage to LandingPage,
as a transiction. Instead when the user resumes the app
directly the LandingPage must be showed him.