Added InterProcessCommunication when trying to launch a second instance of the application.
BEHAVIOUR:
When the app starts-up everythin normal.
When a second instance of the app is launched there are two cases:
- the app has been launched from the Icon
- the app has been launched from contextual menu

So, in the first case, the MainWindow is Activated, and popped-up
to the user. Instead when the app is launched by the contextualMenu the arguments are passed to the active process
and the application is triggered to manage the contextualMenu case.
In faaaact, in this mode, the app will load the BrowsePage and the file transfer overview will be filled with the data passed, ready to selecte an user and start the transfer.